Added Mock character movement
