+++ Assets/_BForBoss/Tests/Mocks/MockCharacterMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4279ed7d00ab41f439edc7cd77c7a050
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
